Datastore/requirements.txt
google-cloud-datastore~=2.15.2